requirements.txt
python-docx
pyyaml